+++ test/e2e/tests/metrics/token-detection-metrics.spec.js
@@ -99,7 +99,6 @@ describe('Token detection event @no-mmi', function () {
           chain_id: '0x539',
           environment_type: 'fullscreen',
           is_profile_syncing_enabled: null,
-          is_signed_in: false,
         });
         assert.deepStrictEqual(events[2].properties, {
           token_detection_enabled: true,