+++ ui/components/app/modals/turn-on-metamask-notifications/turn-on-metamask-notifications.tsx
@@ -61,12 +61,11 @@ export default function TurnOnMetamaskNotifications() {
     setButtonState(true);
     await createNotifications();
     trackEvent({
-      category: MetaMetricsEventCategory.NotificationInteraction,
-      event: MetaMetricsEventName.NotificationMenuOpened,
+      category: MetaMetricsEventCategory.NotificationsActivationFlow,
+      event: MetaMetricsEventName.NotificationsActivated,
       properties: {
-        is_profile_syncing_enabled: isProfileSyncingEnabled,
-        is_notifications_enabled: isNotificationEnabled,
-        action_type: 'enabled',
+        is_profile_syncing_enabled: true,
+        action_type: 'activated',
       },
     });
   };
@@ -74,11 +73,10 @@ export default function TurnOnMetamaskNotifications() {
   const handleHideModal = () => {
     hideModal();
     trackEvent({
-      category: MetaMetricsEventCategory.NotificationInteraction,
-      event: MetaMetricsEventName.NotificationMenuOpened,
+      category: MetaMetricsEventCategory.NotificationsActivationFlow,
+      event: MetaMetricsEventName.NotificationsActivated,
       properties: {
         is_profile_syncing_enabled: isProfileSyncingEnabled,
-        is_notifications_enabled: isNotificationEnabled,
         action_type: 'dismissed',
       },
     });