+++ ui/components/app/wallet-overview/coin-overview.tsx
@@ -4,7 +4,7 @@ import classnames from 'classnames';
 import { zeroAddress } from 'ethereumjs-util';
 import { CaipChainId } from '@metamask/utils';
 import type { Hex } from '@metamask/utils';
-import { Icon, IconName } from '../../component-library';
+import { Icon, IconName, IconSize } from '../../component-library';
 import { IconColor } from '../../../helpers/constants/design-system';
 import { MetaMetricsContext } from '../../../contexts/metametrics';
 import {
@@ -141,14 +141,6 @@ export const CoinOverview = ({
           disabled={!balanceIsCached}
         >
           <div className={`${classPrefix}-overview__balance`}>
-            <div
-              onClick={handlePortfolioOnClick}
-              className="wallet-overview__portfolio_button"
-              data-testid="portfolio-link"
-            >
-              {t('portfolio')}
-              <Icon name={IconName.Diagram} color={IconColor.primaryDefault} />
-            </div>
             <div className={`${classPrefix}-overview__primary-container`}>
               {balanceToDisplay ? (
                 <UserPreferencedCurrencyDisplay
@@ -165,7 +157,7 @@ export const CoinOverview = ({
                   type={PRIMARY}
                   ethNumberOfDecimals={4}
                   hideTitle
-                  withCheckShowNativeToken
+                  shouldCheckShowNativeToken
                   isAggregatedFiatOverviewBalance={
                     !showNativeTokenAsMainBalance
                   }
@@ -178,6 +170,20 @@ export const CoinOverview = ({
                   *
                 </span>
               )}
+            </div>
+            <div className="wallet-overview__currency-wrapper">
+              <div
+                onClick={handlePortfolioOnClick}
+                className="wallet-overview__portfolio_button"
+                data-testid="portfolio-link"
+              >
+                {t('portfolio')}
+                <Icon
+                  size={IconSize.Sm}
+                  name={IconName.Export}
+                  color={IconColor.primaryDefault}
+                />
+              </div>
             </div>
             {renderPercentageAndAmountChange()}
           </div>