+++ ui/hooks/useCurrencyDisplay.js
@@ -8,7 +8,10 @@ import {
   getMultichainConversionRate,
 } from '../selectors/multichain';
 
-import { getValueFromWeiHex } from '../../shared/modules/conversion.utils';
+import {
+  decimalToHex,
+  getValueFromWeiHex,
+} from '../../shared/modules/conversion.utils';
 import { TEST_NETWORK_TICKER_MAP } from '../../shared/constants/network';
 import { Numeric } from '../../shared/modules/Numeric';
 import { EtherDenomination } from '../../shared/constants/common';
@@ -175,7 +178,7 @@ export function useCurrencyDisplay(
     }
 
     if (isAggregatedFiatOverviewBalance) {
-      return inputValue;
+      return formatCurrency(inputValue, currency);
     }
 
     return formatEthCurrencyDisplay({