+++ ui/pages/confirm-encryption-public-key/confirm-encryption-public-key.container.js
@@ -11,7 +11,6 @@ import {
 import {
   unconfirmedTransactionsListSelector,
   getTargetAccountWithSendEtherInfo,
-  getCurrentCurrency,
 } from '../../selectors';
 
 import { clearConfirmTransaction } from '../../ducks/confirm-transaction/confirm-transaction.duck';
@@ -19,7 +18,6 @@ import { getMostRecentOverviewPage } from '../../ducks/history/history';
 import { getNativeCurrency } from '../../ducks/metamask/metamask';
 import ConfirmEncryptionPublicKey from './confirm-encryption-public-key.component';
 
-// Removing usage of useNativeCurrencyAsPrimaryCurrency because component ConfirmEncryptionPublicKey is not
 function mapStateToProps(state) {
   const {
     metamask: { subjectMetadata = {} },
@@ -40,10 +38,8 @@ function mapStateToProps(state) {
     fromAccount,
     requester: null,
     requesterAddress: null,
-    conversionRate: null,
     mostRecentOverviewPage: getMostRecentOverviewPage(state),
     nativeCurrency: getNativeCurrency(state),
-    currentCurrency: getCurrentCurrency(state),
   };
 }