+++ ui/pages/confirmations/components/confirm/footer/footer.tsx
@@ -1,4 +1,6 @@
+///: BEGIN:ONLY_INCLUDE_IF(build-main,build-beta,build-flask)
 import { TransactionMeta } from '@metamask/transaction-controller';
+///: END:ONLY_INCLUDE_IF
 import { ethErrors, serializeError } from 'eth-rpc-errors';
 import React, { useCallback, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
@@ -22,10 +24,19 @@ import useAlerts from '../../../../../hooks/useAlerts';
 import {
   rejectPendingApproval,
   resolvePendingApproval,
+  ///: BEGIN:ONLY_INCLUDE_IF(build-main,build-beta,build-flask)
   updateAndApproveTx,
+  ///: END:ONLY_INCLUDE_IF
 } from '../../../../../store/actions';
 import { confirmSelector } from '../../../selectors';
-import { REDESIGN_DEV_TRANSACTION_TYPES } from '../../../utils';
+import { selectUseTransactionSimulations } from '../../../selectors/preferences';
+
+import {
+  isPermitSignatureRequest,
+  isSIWESignatureRequest,
+  REDESIGN_DEV_TRANSACTION_TYPES,
+} from '../../../utils';
+import { useConfirmContext } from '../../../context/confirm';
 import { getConfirmationSender } from '../utils';
 import { MetaMetricsEventLocation } from '../../../../../../shared/constants/metametrics';
 
@@ -104,12 +115,17 @@ const Footer = () => {
   const t = useI18nContext();
   const confirm = useSelector(confirmSelector);
   const customNonceValue = useSelector(getCustomNonceValue);
+  const useTransactionSimulations = useSelector(
+    selectUseTransactionSimulations,
+  );
 
-  const { currentConfirmation, isScrollToBottomNeeded } = confirm;
+  const { currentConfirmation } = useConfirmContext();
+  const { isScrollToBottomCompleted } = confirm;
   const { from } = getConfirmationSender(currentConfirmation);
 
   ///: BEGIN:ONLY_INCLUDE_IF(build-mmi)
-  const { mmiOnSignCallback, mmiSubmitDisabled } = useMMIConfirmations();
+  const { mmiOnTransactionCallback, mmiOnSignCallback, mmiSubmitDisabled } =
+    useMMIConfirmations();
   ///: END:ONLY_INCLUDE_IF
 
   const hardwareWalletRequiresConnection = useSelector((state) => {
@@ -118,6 +134,17 @@ const Footer = () => {
     }
     return false;
   });
+
+  const isSIWE = isSIWESignatureRequest(currentConfirmation);
+  const isPermit = isPermitSignatureRequest(currentConfirmation);
+  const isPermitSimulationShown = isPermit && useTransactionSimulations;
+
+  const isConfirmDisabled =
+    (!isScrollToBottomCompleted && !isSIWE && !isPermitSimulationShown) ||
+    ///: BEGIN:ONLY_INCLUDE_IF(build-mmi)
+    mmiSubmitDisabled ||
+    ///: END:ONLY_INCLUDE_IF
+    hardwareWalletRequiresConnection;
 
   const onCancel = useCallback(
     ({ location }: { location?: MetaMetricsEventLocation }) => {
@@ -144,6 +171,7 @@ const Footer = () => {
       (type) => type === currentConfirmation?.type,
     );
     if (isTransactionConfirmation) {
+      ///: BEGIN:ONLY_INCLUDE_IF(build-main,build-beta,build-flask)
       const mergeTxDataWithNonce = (transactionData: TransactionMeta) =>
         customNonceValue
           ? { ...transactionData, customNonceValue }
@@ -152,8 +180,15 @@ const Footer = () => {
       const updatedTx = mergeTxDataWithNonce(
         currentConfirmation as TransactionMeta,
       );
+      ///: END:ONLY_INCLUDE_IF
 
+      ///: BEGIN:ONLY_INCLUDE_IF(build-mmi)
+      mmiOnTransactionCallback();
+      ///: END:ONLY_INCLUDE_IF
+
+      ///: BEGIN:ONLY_INCLUDE_IF(build-main,build-beta,build-flask)
       dispatch(updateAndApproveTx(updatedTx, true, ''));
+      ///: END:ONLY_INCLUDE_IF
     } else {
       dispatch(resolvePendingApproval(currentConfirmation.id, undefined));
 
@@ -181,13 +216,7 @@ const Footer = () => {
       <ConfirmButton
         alertOwnerId={currentConfirmation?.id}
         onSubmit={() => onSubmit()}
-        disabled={
-          ///: BEGIN:ONLY_INCLUDE_IF(build-mmi)
-          mmiSubmitDisabled ||
-          ///: END:ONLY_INCLUDE_IF
-          isScrollToBottomNeeded ||
-          hardwareWalletRequiresConnection
-        }
+        disabled={isConfirmDisabled}
         onCancel={onCancel}
       />
     </PageFooter>